Add a backup destination for barn-agent.
Barn currently has a single master destination. This change introduces the
concept of a primary and a secondary destination. When a master is
down the agent will wait a period (currently temporarily hard-coded to 10 minutes)
and then fail over to backup. It will retry master every 10 minutes from
then on.

This change was done because under the current setup barn-agent
has no control over when log files get deleted and the log deletion
time can be as low as 20 minutes which is very short period for a
replacement machine to be brought up.