Report number of files shipped.
Barn-agent reports number of 'files to ship' which when working normally
can be used as a metric of 'files shipped'. However when either (a)
destination address stops working, or (b), the local process starts
emitting lots of log files then the 'files to ship' goes up. Adding
this metric makes it possible to distinguish does failure modes.