Add Easylogging++ as logging library for barn-agent.
As barn-agent logs get mixed together it can be hard to distinguish one
agent from another. Need to prefix log messages with the service_name that
the agent is reporting.
It appears Easylogging++ (http://easylogging.org/) is the simplest logging
library to solve this.

Also moved barn_conf to params.h where it seems more natural and stops the
circular dependence of barn-agent.h with other header files.